% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_airway}
\alias{example_airway}
\title{Example airway dataset}
\format{
An ExpressionSet object
}
\source{
airway package
}
\usage{
example_airway
}
\description{
A small example dataset derived from the airway package.
This dataset is used in the examples to avoid downloading large data during R CMD check.
}
\keyword{datasets}
