% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{expression_summary}
\alias{expression_summary}
\title{Summary of the expression data}
\usage{
expression_summary(expression_set)
}
\arguments{
\item{expression_set}{An expression set object}
}
\value{
Invisible NULL, prints the summary of the expression data to console
}
\description{
Summary of the expression data
}
\examples{
data(example_airway)
expression_set <- example_airway
expression_summary(expression_set)
}
