% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{filter_low_exp}
\alias{filter_low_exp}
\title{Normalize the data, we want to filter low expression genes from the ExpressionSet}
\usage{
filter_low_exp(
  expression_set,
  min_count = 10,
  min_samples = 2,
  already_log = FALSE
)
}
\arguments{
\item{expression_set}{Expression Set object}

\item{min_count}{minimum count threshold, default is set to 10}

\item{min_samples}{minimum number of samples that a gene have to be expressed in, default is set to 2}

\item{already_log}{Logical, whether the data is already log-transformed. Default is FALSE.}
}
\value{
a filtered expression set object without the lowly expressed genes
}
\description{
Normalize the data, we want to filter low expression genes from the ExpressionSet
}
\examples{
\dontrun{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
exp_set_filtered <- filter_low_exp(expression_set, already_log = FALSE)
}
}
