% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{go_enrich}
\alias{go_enrich}
\title{Performs a GO enrichment analysis}
\usage{
go_enrich(gene_list, OrgDb = "org.Hs.eg.db", keyType = "ENSEMBL")
}
\arguments{
\item{gene_list}{A list of genes}

\item{OrgDb}{OrgDb database to use for annotation, default is "org.Hs.eg.db"}

\item{keyType}{Type of gene ID, default is "ENSEMBL"}
}
\value{
A GO enrichment object
}
\description{
Performs a GO enrichment analysis
}
\examples{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
km <- kmeans_clust(expression_set, 5)
gene_list <- names(km$cluster[km$cluster == 1])
go_enrich(gene_list)
}
