% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{go_gse}
\alias{go_gse}
\title{Performs Gene Set Enrichment Analysis}
\usage{
go_gse(
  expression_set,
  condition_col,
  reference_level,
  OrgDb = "org.Hs.eg.db",
  keyType = "ENSEMBL"
)
}
\arguments{
\item{expression_set}{An expression set object}

\item{condition_col}{The column in the phenoData will be used for the differential expression}

\item{reference_level}{The reference level for the condition}

\item{OrgDb}{OrgDb database to use for annotation, default is "org.Hs.eg.db"}

\item{keyType}{Type of gene ID, default is "ENSEMBL"}
}
\value{
A GSEA enrichment object
}
\description{
Performs Gene Set Enrichment Analysis
}
\examples{
\dontrun{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
exp_norm <- quantile_norm(log_transform(filter_low_exp(expression_set)))
go_gse(exp_norm, condition_col = "dex", reference_level = "untrt")
}
}
