% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{hierarchical_clust}
\alias{hierarchical_clust}
\title{Performs a hierarchical clustering on the expression data}
\usage{
hierarchical_clust(expression_set, method = "complete")
}
\arguments{
\item{expression_set}{An expression set object}

\item{method}{The hierarchical clustering method to be used, set to complete as default}
}
\value{
A hierarchical clustering object
}
\description{
Performs a hierarchical clustering on the expression data
}
\examples{
\dontrun{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
hc_tree <- hierarchical_clust(expression_set, "complete")
hc_clust <- cutree(hc_tree, k = 5)
}
}
