% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{kegg_enrich}
\alias{kegg_enrich}
\title{Performs a KEGG enrichment analysis}
\usage{
kegg_enrich(
  gene_list,
  OrgDb = "org.Hs.eg.db",
  organism = "hsa",
  keyType = "ENSEMBL"
)
}
\arguments{
\item{gene_list}{A list of genes}

\item{OrgDb}{OrgDb database to use for annotation, default is "org.Hs.eg.db"}

\item{organism}{KEGG organism code, default is "hsa"}

\item{keyType}{Type of gene ID, default is "ENSEMBL"}
}
\value{
A KEGG enrichment object
}
\description{
Performs a KEGG enrichment analysis
}
\examples{
\dontrun{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
km <- kmeans_clust(expression_set, 5)
gene_list <- names(km$cluster[km$cluster == 1])
kegg_enrich(gene_list)
}
}
