% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{kmeans_clust}
\alias{kmeans_clust}
\title{This script performs a Kmeans clustering on the expression data}
\usage{
kmeans_clust(expression_set, k_clusters = 5, seed = 42)
}
\arguments{
\item{expression_set}{An expression set object}

\item{k_clusters}{Number of clusters, default is set to 5}

\item{seed}{Seed for reproducibility, default is set to 42}
}
\value{
A kmeans object
}
\description{
This script performs a Kmeans clustering on the expression data
}
\examples{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
km_result <- kmeans_clust(expression_set, 5)
km_clust <- km_result$cluster
}
