% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadfromCSV}
\alias{loadfromCSV}
\title{This function will load gene expression data from a CSV file and read in ExpressionSet}
\usage{
loadfromCSV(file)
}
\arguments{
\item{file}{Represents a path to a CSV file with genes as rows and samples as columns}
}
\value{
An ExpressionSet object
}
\description{
This function will load gene expression data from a CSV file and read in ExpressionSet
}
\examples{
expression_set <- loadfromCSV("path/to/data.csv")
}
