% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadfromSumE}
\alias{loadfromSumE}
\title{This function can read a summarized experiment object (like we are using as an example, the airway dataset) and convert to expression set}
\usage{
loadfromSumE(SumE)
}
\arguments{
\item{SumE}{A SummarizedExperiment object}
}
\value{
An ExpressionSet object
}
\description{
This function can read a summarized experiment object (like we are using as an example, the airway dataset) and convert to expression set
}
\examples{
data(example_airway)
expression_set <- example_airway
}
