% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{log_transform}
\alias{log_transform}
\title{Next step is to logtransform the data}
\usage{
log_transform(expression_set, pseudo_count = 1)
}
\arguments{
\item{expression_set}{Expression Set object}

\item{pseudo_count}{to avoid log(0)}
}
\value{
a log transformed expression set object
}
\description{
Next step is to logtransform the data
}
\examples{
\dontrun{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
exp_set_filtered <- filter_low_exp(expression_set)
exp_set_log <- log_transform(exp_set_filtered)
}
}
