% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{optimal_k}
\alias{optimal_k}
\title{Plot the Elbow Method to find the optimal number of clusters}
\usage{
optimal_k(expression_set, max_k = 10)
}
\arguments{
\item{expression_set}{An expression set object}

\item{max_k}{The maximum number of clusters to test, default is 10}
}
\value{
A ggplot object showing the elbow plot
}
\description{
Plot the Elbow Method to find the optimal number of clusters
}
\examples{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
optimal_k(expression_set)
}
