% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{A function to plot the expression distribution using boxplots per sample}
\usage{
plot_boxplot(
  expression_set,
  title = "Boxplot of Expression Data",
  log_scale = FALSE
)
}
\arguments{
\item{expression_set}{An expression set object}

\item{title}{Plot title for the boxplot}

\item{log_scale}{Logical, whether to apply a log2 transformation before plotting. Default is FALSE.}
}
\value{
a ggplot2 boxplot of the expression distribution
}
\description{
A function to plot the expression distribution using boxplots per sample
}
\examples{
data(example_airway)
expression_set <- example_airway
expression_set <- filter_low_exp(expression_set)
plot_boxplot(expression_set, log_scale = TRUE)
}
