% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_den}
\alias{plot_den}
\title{Plot dendrogram}
\usage{
plot_den(hc)
}
\arguments{
\item{hc}{An hclust object}
}
\value{
plots the dendrogram
}
\description{
Plot dendrogram
}
\examples{
\dontrun{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
exp_set_filtered <- filter_low_exp(expression_set)
exp_set_log <- log_transform(exp_set_filtered)
exp_set_norm <- quantile_norm(exp_set_log)
hc <- hierarchical_clust(exp_set_norm)
plot_den(hc$hclust)
}
}
