% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_distr}
\alias{plot_distr}
\title{A function to plot the distribution of expression values per sample}
\usage{
plot_distr(expression_set, title = "Distribution of Expression Data")
}
\arguments{
\item{expression_set}{An expression set object}

\item{title}{Plot title for the distribution plot}
}
\value{
a ggplot2 plot of the expression distribution
}
\description{
A function to plot the distribution of expression values per sample
}
\examples{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
expression_set <- filter_low_exp(expression_set)
expression_set <- log_transform(expression_set)
plot_distr(expression_set)
}
