% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot the enrichment results}
\usage{
plot_enrichment(enrichment_result, top_n = 10)
}
\arguments{
\item{enrichment_result}{An enrichment result object}

\item{top_n}{Number of top enriched terms to plot}
}
\value{
A plot of the enrichment results
}
\description{
Plot the enrichment results
}
\examples{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
km <- kmeans_clust(expression_set, 5)
gene_list <- names(km$cluster[km$cluster == 1])
go_result <- go_enrich(gene_list)
plot_enrichment(go_result)
}
