% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap using pheatmap}
\usage{
plot_heatmap(
  expression_set,
  show_rownames = FALSE,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  scale = "row",
  top_n_genes = 1000
)
}
\arguments{
\item{expression_set}{An expression set object}

\item{show_rownames}{logical, whether to show row names}

\item{cluster_rows}{logical, whether to cluster rows}

\item{cluster_cols}{logical, whether to cluster columns}

\item{scale}{scale, whether to scale the data}

\item{top_n_genes}{Integer, the number of top variable genes to plot if the dataset is large. Default is 1000.}
}
\value{
a heatmap of the gene expression
}
\description{
Plot heatmap using pheatmap
}
\examples{
\donttest{
data(example_airway)
expression_set <- example_airway
exp_set_filtered <- filter_low_exp(expression_set)
exp_set_log <- log_transform(exp_set_filtered)
exp_set_norm <- quantile_norm(exp_set_log)
plot_heatmap(exp_set_norm)
}
}
