% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA plot}
\usage{
plot_PCA(expression_set, top_genes_pca = 500)
}
\arguments{
\item{expression_set}{An expression set object}

\item{top_genes_pca}{The number of top variable genes to use for PCA. Default is 500.}
}
\value{
a PCA plot of the gene expression
}
\description{
Plot PCA plot
}
\examples{
data(example_airway, package = "GeneExpressionAnalysis")
expression_set <- example_airway
exp_set_filtered <- filter_low_exp(expression_set)
exp_set_log <- log_transform(exp_set_filtered)
exp_set_norm <- quantile_norm(exp_set_log)
plot_PCA(exp_set_norm)
}
