% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_sample_den}
\alias{plot_sample_den}
\title{Plot dendrogram of samples for better interpretation of the data}
\usage{
plot_sample_den(expression_set, method = "complete")
}
\arguments{
\item{expression_set}{An expression set object}

\item{method}{The hierarchical clustering method to be used, default is "complete"}
}
\value{
plots the dendrogram
}
\description{
Plot dendrogram of samples for better interpretation of the data
}
\examples{
\dontrun{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
exp_set_filtered <- filter_low_exp(expression_set)
exp_set_log <- log_transform(exp_set_filtered)
exp_set_norm <- quantile_norm(exp_set_log)
plot_sample_den(exp_set_norm)
}
}
