% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outlier genes from the expression set}
\usage{
remove_outliers(expression_set, threshold = 3)
}
\arguments{
\item{expression_set}{An expression set object}

\item{threshold}{Tolerated standard deviations from the mean}
}
\value{
A filtered expression set object with outlier genes removed
}
\description{
Remove outlier genes from the expression set
}
\examples{
data(example_airway, package = "GeneExpressionAnalysis")
expression_set <- example_airway
expression_set <- filter_low_exp(expression_set)
exp_clean <- remove_outliers(expression_set, threshold = 3)
}
