% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_full_pipeline}
\alias{run_full_pipeline}
\title{The function runs the full pipeline}
\usage{
run_full_pipeline(
  expression_set,
  k_clusters = 5,
  method = "complete",
  top_n = 10,
  enrich_cluster = 1,
  condition_col = "dex",
  reference_level = "untrt",
  OrgDb = "org.Hs.eg.db",
  organism = "hsa",
  keyType = "ENSEMBL",
  already_log = FALSE,
  min_count = 10,
  min_samples = 2,
  top_n_genes = 1000,
  top_genes_pca = 500,
  threshold = 3
)
}
\arguments{
\item{expression_set}{An expression set object}

\item{k_clusters}{Number of clusters, default is set to 5}

\item{method}{The hierarchical clustering method to be used, default is set to "complete"}

\item{top_n}{Number of top enriched terms to plot, default is set to 10}

\item{enrich_cluster}{The cluster to be used for the enrichment analysis, default is set to 1}

\item{condition_col}{The column in the phenoData to be used for the enrichment analysis, default is set to "dex"}

\item{reference_level}{The reference level for the enrichment analysis, default is set to "untrt"}

\item{OrgDb}{The organism database to be used for the enrichment analysis, default is set to "org.Hs.eg.db"}

\item{organism}{The organism to be used for the enrichment analysis, default is set to "hsa"}

\item{keyType}{Type of gene ID, default is "ENSEMBL"}

\item{already_log}{Logical, whether the data is already log-transformed. Default is FALSE.}

\item{min_count}{minimum count threshold for filtering, default is set to 10}

\item{min_samples}{minimum number of samples for filtering, default is set to 2}

\item{top_n_genes}{Number of genes to show in heatmap, default is 1000}

\item{top_genes_pca}{Number of genes to use for PCA, default is 500}

\item{threshold}{Standard deviation threshold for outlier detection, default is 3}
}
\value{
A list containing:
\itemize{
\item expression_set: The processed ExpressionSet
\item kmeans: The kmeans clustering result
\item hclust: The hierarchical clustering result (list with hclust and dist)
\item go: GO enrichment results
\item kegg: KEGG enrichment results
\item gsea: GSEA enrichment results
}
}
\description{
The function runs the full pipeline
}
\examples{
\dontrun{
data(example_airway, package = "GeneExpressionAnalysis")
expression_set <- example_airway
run_full_pipeline(expression_set)
}
}
