% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{sample_outlier_detection}
\alias{sample_outlier_detection}
\title{This function will detect high variance samples}
\usage{
sample_outlier_detection(expression_set, threshold = 3)
}
\arguments{
\item{expression_set}{An expression set object}

\item{threshold}{Tolerated standard deviations from the mean}
}
\value{
Outlier samples
}
\description{
This function will detect high variance samples
}
\examples{
data(example_airway)
expression_set <- example_airway
expression_set <- filter_low_exp(expression_set)
expression_set <- log_transform(expression_set)
sample_outlier_detection(expression_set, threshold = 3)
}
