% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{log_transform}
\alias{log_transform}
\title{This function logtransform the data}
\usage{
log_transform(expression_set, pseudo_count = 1)
}
\arguments{
\item{expression_set}{Expression Set object}

\item{pseudo_count}{to avoid log(0)}
}
\value{
a log transformed expression set object
}
\description{
This function logtransform the data
}
\examples{
data(example_airway)
expression_set <- example_airway
exp_set_filtered <- filter_low_exp(expression_set)
exp_set_log <- log_transform(exp_set_filtered)
}
