% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_distr}
\alias{plot_distr}
\title{This function plots the distribution of expression values per sample}
\usage{
plot_distr(expression_set, title = "Distribution of Expression Data")
}
\arguments{
\item{expression_set}{An expression set object}

\item{title}{Plot title for the distribution plot}
}
\value{
a ggplot2 plot of the expression distribution
}
\description{
This function plots the distribution of expression values per sample
}
\examples{
data(example_airway, package = "GeneExpressionAnalysis")
expression_set <- example_airway
expression_set <- filter_low_exp(expression_set)
expression_set <- log_transform(expression_set)
plot_distr(expression_set)
}
