% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{quantile_norm}
\alias{quantile_norm}
\title{This function performs a quantile normalization}
\usage{
quantile_norm(expression_set)
}
\arguments{
\item{expression_set}{Expression Set object}
}
\value{
a quantile normalized expression set object
}
\description{
This function performs a quantile normalization
}
\examples{
\dontrun{
library(airway)
data(airway)
expression_set <- loadfromSumE(airway)
exp_set_filtered <- filter_low_exp(expression_set)
exp_set_log <- log_transform(exp_set_filtered)
exp_set_norm <- quantile_norm(exp_set_log)
}
}
